requirements.txt
algoliasearch-django>=2.0,<3.0
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
Markdown==3.4.1
psycopg2==2.9.3
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11